+++ include/kern/drivers/mems.h
@@ -6,7 +6,7 @@
 #include <stdint.h>
 
 size_t mem_get_num(void);
-long mem_get_addr(unsigned i, uint64_t *addr);
-long mem_get_size(unsigned i, uint64_t *size);
+long mem_get_addr(unsigned i, uintptr_t *addr);
+long mem_get_size(unsigned i, uintmax_t *size);
 
 #endif
