+++ include/kern/mm/pmm.h
@@ -13,8 +13,8 @@ unsigned long mm_get_freemem_base(void);
 void *palloc(size_t size, size_t align);
 
 void pmm_init(void);
-long pa2frame(unsigned long addr, struct phy_frame **frame) __attribute__((warn_unused_result));
-long frame2pa(struct phy_frame *frame, unsigned long *addr) __attribute__((warn_unused_result));
+long pa2frame(uintptr_t addr, struct phy_frame **frame) __attribute__((warn_unused_result));
+long frame2pa(struct phy_frame *frame, uintptr_t *addr) __attribute__((warn_unused_result));
 long phy_frame_alloc(struct phy_frame **frame) __attribute__((warn_unused_result));
 long phy_frame_ref_dec(struct phy_frame *frame) __attribute__((warn_unused_result));
 long phy_frame_ref_inc(struct phy_frame *frame) __attribute__((warn_unused_result));
